requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.7
django-cache-url==3.2.3
environs==8.0.0
marshmallow==3.13.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==21.0
Pillow==8.2.0
pycodestyle==2.7.0
pymongo==3.12.0
pyparsing==2.4.7
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0